requirements.txt
# `pip install -r requirements.txt` before you run this package.
matplotlib
pyyaml
pandas
sklearn
seaborn